requirements.txt
fastapi
uvicorn
Jinja2
SQLAlchemy
pytailwindcss
passlib
bcrypt
python-multipart
pyjwt
pytest